fix type

+++ python_variable_annotations/0-add.py
@@ -1,5 +1,5 @@
 #!/usr/bin/env python3
 """""""
-def add(a: int, b: int) -> int:
+def add(a: float, b: float) -> int:
     """""""
     return a + b
